llm/rag_ondemand/docker/requirements.txt
python-dotenv
fastapi
uvicorn
python-multipart
pydantic
boto3
requests
starlette
passlib
bcrypt
python-jose[cryptography]
python-magic
